backend/requirements.txt
Django>=5.1.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
Pillow>=10.0.0
psycopg2-binary>=2.9.7
python-decouple>=3.8 